% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta_file.R
\name{build_meta_file}
\alias{build_meta_file}
\title{create meta data for new files}
\usage{
build_meta_file(path = ".")
}
\arguments{
\item{path}{The path where the data has to be stored}
}
\value{
a tibble where you have to set the parameters by hand.
}
\description{
create meta data for new files
}
\examples{
\dontrun{
build_meta_file()
}
}
