% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{function for creating a rmd file and copying it to the webpage folder}
\usage{
create_report(path_name, out_path, cats)
}
\arguments{
\item{path_name}{name of the folder}

\item{out_path}{The path where the data lays}

\item{cats}{add categories that will show up after render}
}
\description{
function for creating a rmd file and copying it to the webpage folder
}
\examples{
\dontrun{
create_report()
}
}
