% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_name.R
\name{data_name}
\alias{data_name}
\title{create name of data files}
\usage{
data_name()
}
\value{
a data name in a given format
}
\description{
create name of data files
}
\examples{
\dontrun{
data_name()
}
}
