% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta_file.R
\name{empty_as_na}
\alias{empty_as_na}
\title{function for empty cols}
\usage{
empty_as_na(x)
}
\arguments{
\item{x}{input}
}
\value{
if x is "" then it will be converted to NA
}
\description{
function for empty cols
}
\examples{
\dontrun{
fun_date("")
}
}
