% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta_file.R
\name{fun_date}
\alias{fun_date}
\title{function for dates}
\usage{
fun_date(x)
}
\arguments{
\item{x}{input}
}
\value{
returns a date
}
\description{
function for dates
}
\examples{
\dontrun{
fun_date("2017-01-01")
}
}
