% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fun_num}
\alias{fun_num}
\title{function for numeric}
\usage{
fun_num(x)
}
\arguments{
\item{x}{input}
}
\value{
returns a number
}
\description{
function for numeric
}
\examples{
\dontrun{
fun_date("1")
}
}
