% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dropbox_path.R
\name{get_dropbox_path}
\alias{get_dropbox_path}
\title{get the path to our dropbox}
\usage{
get_dropbox_path()
}
\value{
the path to the dropbox
}
\description{
get the path to our dropbox
}
\examples{
\dontrun{
get_dropbox_path()
}
}
