% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta_file.R
\name{get_license}
\alias{get_license}
\title{function for getting license of the data}
\usage{
get_license(source, year = lubridate::year(Sys.Date()))
}
\arguments{
\item{source}{source name}

\item{year}{year of downloaded data}
}
\value{
a character string with the respective license for citation
}
\description{
function for getting license of the data
}
\examples{
\dontrun{
render_geodata()
}
}
