% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_path.R
\name{get_user_path}
\alias{get_user_path}
\title{get the path of the user}
\usage{
get_user_path()
}
\value{
the path of the user
}
\description{
get the path of the user
}
\examples{
\dontrun{
get_user_path()
}
}
