% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeodataapp.R
\name{getgeodataApp}
\alias{getgeodataApp}
\title{access shiny app for browse geodata on server and may download them}
\usage{
getgeodataApp()
}
\value{
an app
}
\description{
access shiny app for browse geodata on server and may download them
}
\examples{
\dontrun{
getgeodataApp()
}
}
