% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantitative_map.R
\name{plot_quantitative_map}
\alias{plot_quantitative_map}
\title{plot quantitative maps}
\usage{
plot_quantitative_map(tif)
}
\arguments{
\item{tif}{the raster that should be plotted}
}
\value{
a plot
}
\description{
plot quantitative maps
}
\examples{
\dontrun{
plot_quantitative_map()
}
}
