% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_geodata.R
\name{render_geodata}
\alias{render_geodata}
\title{function for rendering an markdown to get an html of available geodata}
\usage{
render_geodata(path)
}
\arguments{
\item{path}{The path where the data has to be stored}
}
\value{
a tibble where you have to set the parameters by hand.
}
\description{
function for rendering an markdown to get an html of available geodata
}
\examples{
\dontrun{
render_geodata()
}
}
