% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_overview.R
\name{geo_overview}
\alias{geo_overview}
\title{show overview of geodata}
\usage{
geo_overview(path_to_cloud)
}
\arguments{
\item{path_to_cloud}{set the path to the cloud}
}
\value{
a list with folder names
}
\description{
show overview of geodata
}
\examples{
\dontrun{
geo_overview()
}
}
