% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_to_distill.R
\name{move_to_distill}
\alias{move_to_distill}
\title{function for copying the geodata template, adding the paths and rename it to the corresponding data}
\usage{
move_to_distill(
  path,
  to_path = "C:/Users/wenzler/Documents/GitHub/d6geodatabase"
)
}
\arguments{
\item{path}{The path where the data lays}

\item{to_path}{where to move. Fixed!}
}
\description{
function for copying the geodata template, adding the paths and rename it to the corresponding data
}
\examples{
\dontrun{
move_to_distill()
}
}
