% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binary_map.R
\name{plot_binary_map}
\alias{plot_binary_map}
\title{plot binary maps}
\usage{
plot_binary_map(tif)
}
\arguments{
\item{tif}{the raster that should be plotted}
}
\value{
a plot
}
\description{
plot binary maps
}
\examples{
\dontrun{
plot_binary_map()
}
}
