% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binary_map.R
\name{plot_binary_map}
\alias{plot_binary_map}
\title{plot binary maps}
\usage{
plot_binary_map(tif, p_base_map)
}
\value{
a plot
}
\description{
plot binary maps
}
\examples{
\dontrun{
d6plot_binary_map()
}
}
