% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qualitative_map.R
\name{plot_qualitative_map}
\alias{plot_qualitative_map}
\title{plot qualitative maps}
\usage{
plot_qualitative_map(tif, p_base_map)
}
\arguments{
\item{tif}{the raster that should be plotted}

\item{p_base_map}{base map}
}
\value{
a plot
}
\description{
plot qualitative maps
}
\examples{
\dontrun{
plot_qualitative_map()
}
}
