% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_raster_berlin.R
\name{temp_ras}
\alias{temp_ras}
\title{get template raster for Berlin}
\usage{
temp_ras(resolution)
}
\arguments{
\item{resolution}{resolution in meters}
}
\value{
a template spatRaster
}
\description{
get template raster for Berlin
}
\examples{
\dontrun{
temp_rast(resolution)
}
}
